wip(ix): refactor choices/console/tabs into seperate classes
